requirements.txt
python-dotenv
requests
groq
PyPDF2
python-dotenv
groq
PyMuPDF
requests
beautifulsoup4
